1.Unify the logic of dedicated and shared KMS AAD 2.Optimize readme documentation and samples.
