+++ examples/src/main/java/com/aliyun/kms/kms20160120/samples/AsymmetricDecryptSample.java
@@ -1,26 +1,24 @@
-package com.aliyun.kms.kms20160120;
+package com.aliyun.kms.kms20160120.samples;
 
 import com.aliyun.kms.kms20160120.Client;
 import com.aliyun.kms.kms20160120.model.KmsConfig;
 import com.aliyun.kms.kms20160120.model.KmsRuntimeOptions;
-import com.aliyun.kms20160120.models.GenerateDataKeyRequest;
-import com.aliyun.kms20160120.models.GenerateDataKeyResponse;
+import com.aliyun.kms20160120.models.AsymmetricDecryptRequest;
+import com.aliyun.kms20160120.models.AsymmetricDecryptResponse;
 import com.aliyun.tea.TeaException;
 
 
-
-public class GenerateDataKeySample {
-    
+public class AsymmetricDecryptSample {
     public static void main(String[] args) throws Exception {
-        newUserGenerateDataKey();
-        beforeMigrateGenerateDataKey();
-        afterMigrateGenerateDataKey();
+        newUserAsymmetricDecrypt();
+        beforeMigrateAsymmetricDecrypt();
+        afterMigrateAsymmetricDecrypt();
     }
 
     /**
      * 新接入用户可以参考此方法调用KMS实例网关的服务。
      */
-    private static void newUserGenerateDataKey() throws Exception {
+    private static void newUserAsymmetricDecrypt() throws Exception {
         //创建kms config并设置相应参数
         com.aliyun.dkms.gcs.openapi.models.Config config
                 = new KmsConfig()
@@ -33,13 +31,13 @@ public class GenerateDataKeySample {
                 //设置client-key密码
                 .setPassword("your-password")
                 //设置dkms域名
-                .setEndpoint("your-dkms-endpoint");
-        // 如需验证服务端证书，这里需要设置为您的服务端证书路径
-        //.setCaFilePath("path/to/yourCaCert")
+                .setEndpoint("your-dkms-endpoint")
+                // 如需验证服务端证书，这里需要设置为您的服务端证书路径
+                .setCaFilePath("path/to/yourCaCert");
         // 或者，设置为您的服务端证书内容
         //.setCa("your-ca-certificate-content"));
         Client client = new Client(config);
-        generateDataKey(client);
+        asymmetricDecrypt(client);
     }
 
     /**
@@ -47,7 +45,7 @@ public class GenerateDataKeySample {
      *
      * @throws Exception
      */
-    private static void beforeMigrateGenerateDataKey() throws Exception {
+    private static void beforeMigrateAsymmetricDecrypt() throws Exception {
         //创建kms config并设置相应参数
         com.aliyun.teaopenapi.models.Config config = new com.aliyun.teaopenapi.models.Config()
                 //设置KMS共享网关的域名
@@ -58,7 +56,7 @@ public class GenerateDataKeySample {
                 .setAccessKeySecret(System.getenv("your-sk-env-name"));
         //创建kms client
         Client client = new Client(config);
-        generateDataKey(client);
+        asymmetricDecrypt(client);
     }
 
     /**
@@ -66,7 +64,7 @@ public class GenerateDataKeySample {
      *
      * @throws Exception
      */
-    private static void afterMigrateGenerateDataKey() throws Exception {
+    private static void afterMigrateAsymmetricDecrypt() throws Exception {
         //创建kms config并设置相应参数
         com.aliyun.teaopenapi.models.Config config = new com.aliyun.teaopenapi.models.Config()                //设置KMS共享网关的域名
                 .setEndpoint("your-kms-endpoint")
@@ -86,38 +84,39 @@ public class GenerateDataKeySample {
                 //设置client-key密码
                 .setPassword("your-password")
                 //设置dkms域名
-                .setEndpoint("your-dkms-endpoint");
-        // 如需验证服务端证书，这里需要设置为您的服务端证书路径
-        //.setCaFilePath("path/to/yourCaCert")
+                .setEndpoint("your-dkms-endpoint")
+                // 如需验证服务端证书，这里需要设置为您的服务端证书路径
+                .setCaFilePath("path/to/yourCaCert");
         // 或者，设置为您的服务端证书内容
         //.setCa("your-ca-certificate-content"));
         //创建kms client
         Client client = new Client(config, kmsConfig);
-        generateDataKey(client);
+        asymmetricDecrypt(client);
     }
 
-
     /**
-     * 生成数据密钥
+     * 非对称加密
      */
-    public static void generateDataKey(Client client) {
+    private static void asymmetricDecrypt(Client client) {
 
         try {
-            GenerateDataKeyRequest request = new GenerateDataKeyRequest();
+            AsymmetricDecryptRequest request = new AsymmetricDecryptRequest();
+            request.setCiphertextBlob("your-ciphertext-blob");
+            request.setAlgorithm("your-algorithm");
             request.setKeyId("your-key-id");
             KmsRuntimeOptions runtimeOptions = new KmsRuntimeOptions();
             //如需忽略SSL证书认证,可打开如下代码并设置ignoreSSLVerifySwitch为true
             //runtimeOptions.ignoreSSL = true;
-            GenerateDataKeyResponse response = client.generateDataKeyWithOptions(request, runtimeOptions);
-            System.out.printf("KeyId: %s%n", response.getBody().getKeyId());
-            System.out.printf("KeyVersionId: %s%n", response.getBody().getKeyVersionId());
-            System.out.printf("CiphertextBlob: %s%n", response.getBody().getCiphertextBlob());
-            System.out.printf("Plaintext: %s%n", response.getBody().getPlaintext());
-            System.out.printf("RequestId: %s%n", response.getBody().getRequestId());
+            AsymmetricDecryptResponse response = client.asymmetricDecryptWithOptions(request, runtimeOptions);
+            System.out.printf("KeyId: %s%n", response.getBody().keyId);
+            System.out.printf("KeyVersionId: %s%n", response.getBody().keyVersionId);
+            System.out.printf("Plaintext: %s%n", response.getBody().plaintext);
         } catch (TeaException e) {
             e.printStackTrace();
+            throw new RuntimeException(e);
         } catch (Exception e) {
             e.printStackTrace();
         }
     }
+
 }
