+++ examples/src/main/java/com/aliyun/kms/kms20160120/samples/AsymmetricEncryptSample.java
@@ -1,26 +1,25 @@
-package com.aliyun.kms.kms20160120;
+package com.aliyun.kms.kms20160120.samples;
 
 import com.aliyun.kms.kms20160120.Client;
 import com.aliyun.kms.kms20160120.model.KmsConfig;
 import com.aliyun.kms.kms20160120.model.KmsRuntimeOptions;
-import com.aliyun.kms20160120.models.GetSecretValueRequest;
-import com.aliyun.kms20160120.models.GetSecretValueResponse;
+import com.aliyun.kms20160120.models.AsymmetricEncryptRequest;
+import com.aliyun.kms20160120.models.AsymmetricEncryptResponse;
 import com.aliyun.tea.TeaException;
-import com.aliyun.teautil.models.RuntimeOptions;
 
 
-public class GetSecretValueSample {
+public class AsymmetricEncryptSample {
 
     public static void main(String[] args) throws Exception {
-        newUserGetSecretValue();
-        beforeMigrateGetSecretValue();
-        afterMigrateGetSecretValue();
+        newUserAsymmetricEncrypt();
+        beforeMigrateAsymmetricEncrypt();
+        afterMigrateAsymmetricEncrypt();
     }
 
     /**
      * 新接入用户可以参考此方法调用KMS实例网关的服务。
      */
-    private static void newUserGetSecretValue() throws Exception {
+    private static void newUserAsymmetricEncrypt() throws Exception {
         //创建kms config并设置相应参数
         com.aliyun.dkms.gcs.openapi.models.Config config
                 = new KmsConfig()
@@ -33,13 +32,13 @@ public class GetSecretValueSample {
                 //设置client-key密码
                 .setPassword("your-password")
                 //设置dkms域名
-                .setEndpoint("your-dkms-endpoint");
-        // 如需验证服务端证书，这里需要设置为您的服务端证书路径
-        //.setCaFilePath("path/to/yourCaCert")
+                .setEndpoint("your-dkms-endpoint")
+                // 如需验证服务端证书，这里需要设置为您的服务端证书路径
+                .setCaFilePath("path/to/yourCaCert");
         // 或者，设置为您的服务端证书内容
         //.setCa("your-ca-certificate-content"));
         Client client = new Client(config);
-        getSecretValue(client);
+        asymmetricEncrypt(client);
     }
 
     /**
@@ -47,7 +46,7 @@ public class GetSecretValueSample {
      *
      * @throws Exception
      */
-    private static void beforeMigrateGetSecretValue() throws Exception {
+    private static void beforeMigrateAsymmetricEncrypt() throws Exception {
         //创建kms config并设置相应参数
         com.aliyun.teaopenapi.models.Config config = new com.aliyun.teaopenapi.models.Config()
                 //设置KMS共享网关的域名
@@ -58,7 +57,7 @@ public class GetSecretValueSample {
                 .setAccessKeySecret(System.getenv("your-sk-env-name"));
         //创建kms client
         Client client = new Client(config);
-        getSecretValue(client);
+        asymmetricEncrypt(client);
     }
 
     /**
@@ -66,7 +65,7 @@ public class GetSecretValueSample {
      *
      * @throws Exception
      */
-    private static void afterMigrateGetSecretValue() throws Exception {
+    private static void afterMigrateAsymmetricEncrypt() throws Exception {
         //创建kms config并设置相应参数
         com.aliyun.teaopenapi.models.Config config = new com.aliyun.teaopenapi.models.Config()                //设置KMS共享网关的域名
                 .setEndpoint("your-kms-endpoint")
@@ -86,31 +85,35 @@ public class GetSecretValueSample {
                 //设置client-key密码
                 .setPassword("your-password")
                 //设置dkms域名
-                .setEndpoint("your-dkms-endpoint");
-        // 如需验证服务端证书，这里需要设置为您的服务端证书路径
-        //.setCaFilePath("path/to/yourCaCert")
+                .setEndpoint("your-dkms-endpoint")
+                // 如需验证服务端证书，这里需要设置为您的服务端证书路径
+                .setCaFilePath("path/to/yourCaCert");
         // 或者，设置为您的服务端证书内容
         //.setCa("your-ca-certificate-content"));
         //创建kms client
         Client client = new Client(config, kmsConfig);
-        getSecretValue(client);
+        asymmetricEncrypt(client);
     }
 
-    public static void getSecretValue(Client client) {
+
+    /**
+     * 非对称加密
+     */
+    public static void asymmetricEncrypt(Client client) {
 
         try {
-            //创建GetSecretValue请求体
-            GetSecretValueRequest request = new GetSecretValueRequest();
-            //设置SecretName参数
-            request.setSecretName("your-secret-name");
-            //创建kms 运行时配置参数
-            RuntimeOptions runtimeOptions = new KmsRuntimeOptions();
+            AsymmetricEncryptRequest request = new AsymmetricEncryptRequest();
+            request.setPlaintext("your-plaintext");
+            request.setKeyId("your-key-id");
+            request.setAlgorithm("your-algorithm");
+            request.setKeyVersionId("your-key-version-id");
+            KmsRuntimeOptions runtimeOptions = new KmsRuntimeOptions();
             //如需忽略SSL证书认证,可打开如下代码并设置ignoreSSLVerifySwitch为true
             //runtimeOptions.ignoreSSL = true;
-            GetSecretValueResponse response = client.getSecretValueWithOptions(request, runtimeOptions);
-            System.out.printf("SecretData: %s%n", response.getBody().getSecretData());
-            System.out.printf("ExtendedConfig: %s%n", response.getBody().getExtendedConfig());
-            System.out.printf("CreateTime: %s%n", response.getBody().getCreateTime());
+            AsymmetricEncryptResponse response = client.asymmetricEncryptWithOptions(request, runtimeOptions);
+            System.out.printf("CiphertextBlob: %s%n", response.getBody().ciphertextBlob);
+            System.out.printf("KeyId: %s%n", response.getBody().keyId);
+            System.out.printf("KeyVersionId: %s%n", response.getBody().keyVersionId);
         } catch (TeaException e) {
             e.printStackTrace();
             throw new RuntimeException(e);
@@ -118,5 +121,4 @@ public class GetSecretValueSample {
             e.printStackTrace();
         }
     }
-
 }
