+++ examples/src/main/java/com/aliyun/kms/kms20160120/samples/DecryptSample.java
@@ -1,25 +1,25 @@
-package com.aliyun.kms.kms20160120;
+package com.aliyun.kms.kms20160120.samples;
 
 import com.aliyun.kms.kms20160120.Client;
 import com.aliyun.kms.kms20160120.model.KmsConfig;
 import com.aliyun.kms.kms20160120.model.KmsRuntimeOptions;
-import com.aliyun.kms20160120.models.GetPublicKeyRequest;
-import com.aliyun.kms20160120.models.GetPublicKeyResponse;
+import com.aliyun.kms20160120.models.DecryptRequest;
+import com.aliyun.kms20160120.models.DecryptResponse;
 import com.aliyun.tea.TeaException;
 
 
-public class GetPublicKeySample {
+public class DecryptSample {
 
     public static void main(String[] args) throws Exception {
-        newUserGetPublicKey();
-        beforeMigrateGetPublicKey();
-        afterMigrateGetPublicKey();
+        newUserDecrypt();
+        beforeMigrateDecrypt();
+        afterMigrateDecrypt();
     }
 
     /**
      * 新接入用户可以参考此方法调用KMS实例网关的服务。
      */
-    private static void newUserGetPublicKey() throws Exception {
+    private static void newUserDecrypt() throws Exception {
         //创建kms config并设置相应参数
         com.aliyun.dkms.gcs.openapi.models.Config config
                 = new KmsConfig()
@@ -32,13 +32,13 @@ public class GetPublicKeySample {
                 //设置client-key密码
                 .setPassword("your-password")
                 //设置dkms域名
-                .setEndpoint("your-dkms-endpoint");
-        // 如需验证服务端证书，这里需要设置为您的服务端证书路径
-        //.setCaFilePath("path/to/yourCaCert")
+                .setEndpoint("your-dkms-endpoint")
+                // 如需验证服务端证书，这里需要设置为您的服务端证书路径
+                .setCaFilePath("path/to/yourCaCert");
         // 或者，设置为您的服务端证书内容
         //.setCa("your-ca-certificate-content"));
         Client client = new Client(config);
-        getPublicKey(client);
+        decrypt(client);
     }
 
     /**
@@ -46,7 +46,7 @@ public class GetPublicKeySample {
      *
      * @throws Exception
      */
-    private static void beforeMigrateGetPublicKey() throws Exception {
+    private static void beforeMigrateDecrypt() throws Exception {
         //创建kms config并设置相应参数
         com.aliyun.teaopenapi.models.Config config = new com.aliyun.teaopenapi.models.Config()
                 //设置KMS共享网关的域名
@@ -57,7 +57,7 @@ public class GetPublicKeySample {
                 .setAccessKeySecret(System.getenv("your-sk-env-name"));
         //创建kms client
         Client client = new Client(config);
-        getPublicKey(client);
+        decrypt(client);
     }
 
     /**
@@ -65,7 +65,7 @@ public class GetPublicKeySample {
      *
      * @throws Exception
      */
-    private static void afterMigrateGetPublicKey() throws Exception {
+    private static void afterMigrateDecrypt() throws Exception {
         //创建kms config并设置相应参数
         com.aliyun.teaopenapi.models.Config config = new com.aliyun.teaopenapi.models.Config()                //设置KMS共享网关的域名
                 .setEndpoint("your-kms-endpoint")
@@ -85,32 +85,32 @@ public class GetPublicKeySample {
                 //设置client-key密码
                 .setPassword("your-password")
                 //设置dkms域名
-                .setEndpoint("your-dkms-endpoint");
-        // 如需验证服务端证书，这里需要设置为您的服务端证书路径
-        //.setCaFilePath("path/to/yourCaCert")
+                .setEndpoint("your-dkms-endpoint")
+                // 如需验证服务端证书，这里需要设置为您的服务端证书路径
+                .setCaFilePath("path/to/yourCaCert");
         // 或者，设置为您的服务端证书内容
         //.setCa("your-ca-certificate-content"));
         //创建kms client
         Client client = new Client(config, kmsConfig);
-        getPublicKey(client);
+        decrypt(client);
     }
 
     /**
-     * 获取公钥
+     * 对称解密
+     *
+     * @param client
      */
-    public static void getPublicKey(Client client) {
-
+    private static void decrypt(Client client) {
         try {
-            GetPublicKeyRequest request = new GetPublicKeyRequest();
-            request.setKeyId("your-key-id");
-            request.setKeyVersionId("your-key-version-id");
+            DecryptRequest request = new DecryptRequest();
+            request.setCiphertextBlob("your-ciphertext-blob");
             KmsRuntimeOptions runtimeOptions = new KmsRuntimeOptions();
             //如需忽略SSL证书认证,可打开如下代码并设置ignoreSSLVerifySwitch为true
             //runtimeOptions.ignoreSSL = true;
-            GetPublicKeyResponse response = client.getPublicKeyWithOptions(request, runtimeOptions);
+            DecryptResponse response = client.decryptWithOptions(request, runtimeOptions);
             System.out.printf("KeyId: %s%n", response.getBody().getKeyId());
             System.out.printf("KeyVersionId: %s%n", response.getBody().getKeyVersionId());
-            System.out.printf("PublicKey: %s%n", response.getBody().getPublicKey());
+            System.out.printf("Plaintext: %s%n", response.getBody().getPlaintext());
             System.out.printf("RequestId: %s%n", response.getBody().getRequestId());
         } catch (TeaException e) {
             e.printStackTrace();
@@ -118,4 +118,5 @@ public class GetPublicKeySample {
             e.printStackTrace();
         }
     }
+
 }
