+++ examples/src/main/java/com/aliyun/kms/kms20160120/samples/EnvelopeDecryptSample.java
@@ -1,25 +1,31 @@
-package com.aliyun.kms.kms20160120;
+package com.aliyun.kms.kms20160120.samples;
 
 import com.aliyun.kms.kms20160120.Client;
 import com.aliyun.kms.kms20160120.model.KmsConfig;
 import com.aliyun.kms.kms20160120.model.KmsRuntimeOptions;
-import com.aliyun.kms20160120.models.AsymmetricVerifyRequest;
-import com.aliyun.kms20160120.models.AsymmetricVerifyResponse;
-import com.aliyun.tea.TeaException;
+import com.aliyun.kms20160120.models.DecryptRequest;
+import com.aliyun.kms20160120.models.DecryptResponse;
+import org.apache.commons.codec.binary.Base64;
 
+import javax.crypto.Cipher;
+import javax.crypto.spec.GCMParameterSpec;
+import javax.crypto.spec.SecretKeySpec;
 
-public class AsymmetricVerifySample {
+public class EnvelopeDecryptSample {
+    private static Base64 base64 = new Base64();
+
+    private static final int GCM_TAG_LENGTH = 16;
 
     public static void main(String[] args) throws Exception {
-        newUserAsymmetricVerify();
-        beforeMigrateAsymmetricVerify();
-        afterMigrateAsymmetricVerify();
+        newUserEnvelopeDecrypt();
+        beforeMigrateEnvelopeDecrypt();
+        afterMigrateEnvelopeDecrypt();
     }
 
     /**
      * 新接入用户可以参考此方法调用KMS实例网关的服务。
      */
-    private static void newUserAsymmetricVerify() throws Exception {
+    private static void newUserEnvelopeDecrypt() throws Exception {
         //创建kms config并设置相应参数
         com.aliyun.dkms.gcs.openapi.models.Config config
                 = new KmsConfig()
@@ -32,13 +38,13 @@ public class AsymmetricVerifySample {
                 //设置client-key密码
                 .setPassword("your-password")
                 //设置dkms域名
-                .setEndpoint("your-dkms-endpoint");
-        // 如需验证服务端证书，这里需要设置为您的服务端证书路径
-        //.setCaFilePath("path/to/yourCaCert")
+                .setEndpoint("your-dkms-endpoint")
+                // 如需验证服务端证书，这里需要设置为您的服务端证书路径
+                .setCaFilePath("path/to/yourCaCert");
         // 或者，设置为您的服务端证书内容
         //.setCa("your-ca-certificate-content"));
         Client client = new Client(config);
-        asymmetricVerify(client);
+        envelopeDecrypt(client);
     }
 
     /**
@@ -46,7 +52,7 @@ public class AsymmetricVerifySample {
      *
      * @throws Exception
      */
-    private static void beforeMigrateAsymmetricVerify() throws Exception {
+    private static void beforeMigrateEnvelopeDecrypt() throws Exception {
         //创建kms config并设置相应参数
         com.aliyun.teaopenapi.models.Config config = new com.aliyun.teaopenapi.models.Config()
                 //设置KMS共享网关的域名
@@ -57,7 +63,7 @@ public class AsymmetricVerifySample {
                 .setAccessKeySecret(System.getenv("your-sk-env-name"));
         //创建kms client
         Client client = new Client(config);
-        asymmetricVerify(client);
+        envelopeDecrypt(client);
     }
 
     /**
@@ -65,7 +71,7 @@ public class AsymmetricVerifySample {
      *
      * @throws Exception
      */
-    private static void afterMigrateAsymmetricVerify() throws Exception {
+    private static void afterMigrateEnvelopeDecrypt() throws Exception {
         //创建kms config并设置相应参数
         com.aliyun.teaopenapi.models.Config config = new com.aliyun.teaopenapi.models.Config()                //设置KMS共享网关的域名
                 .setEndpoint("your-kms-endpoint")
@@ -85,41 +91,53 @@ public class AsymmetricVerifySample {
                 //设置client-key密码
                 .setPassword("your-password")
                 //设置dkms域名
-                .setEndpoint("your-dkms-endpoint");
-        // 如需验证服务端证书，这里需要设置为您的服务端证书路径
-        //.setCaFilePath("path/to/yourCaCert")
+                .setEndpoint("your-dkms-endpoint")
+                // 如需验证服务端证书，这里需要设置为您的服务端证书路径
+                .setCaFilePath("path/to/yourCaCert");
         // 或者，设置为您的服务端证书内容
         //.setCa("your-ca-certificate-content"));
         //创建kms client
         Client client = new Client(config, kmsConfig);
-        asymmetricVerify(client);
+        envelopeDecrypt(client);
     }
 
     /**
-     * 使用KMS API调用DKMS密钥
+     * 信封解密
+     *
+     * @param client
+     * @throws Exception
      */
-    public static void asymmetricVerify(Client client) {
+    private static void envelopeDecrypt(Client client) throws Exception {
+
+        EnvelopeEncryptSample.EnvelopeCipherPersistObject envelopeCipherPersistObject = getEnvelopeCipherPersistObject();
+        String encryptedDataKey = envelopeCipherPersistObject.getEncryptedDataKey();
+        byte[] iv = base64.decode(envelopeCipherPersistObject.getIv());
+        byte[] cipherText = base64.decode(envelopeCipherPersistObject.getCipherText());
 
         try {
-            AsymmetricVerifyRequest request = new AsymmetricVerifyRequest();
-            request.setKeyId("your-key-id");
-            request.setAlgorithm("your-algorithm");
-            request.setDigest("your-digest");
-            request.setValue("your-value");
-            request.setKeyVersionId("your-key-version-id");
+            DecryptRequest decryptRequest = new DecryptRequest();
+            decryptRequest.setCiphertextBlob(encryptedDataKey);
             KmsRuntimeOptions runtimeOptions = new KmsRuntimeOptions();
             //如需忽略SSL证书认证,可打开如下代码并设置ignoreSSLVerifySwitch为true
             //runtimeOptions.ignoreSSL = true;
-            AsymmetricVerifyResponse response = client.asymmetricVerifyWithOptions(request, runtimeOptions);
-            System.out.printf("KeyId: %s%n", response.getBody().getKeyId());
-            System.out.printf("KeyVersionId: %s%n", response.getBody().getKeyVersionId());
-            System.out.printf("Value : %s%n", response.getBody().getValue());
-            System.out.printf("RequestId: %s%n", response.getBody().getRequestId());
-        } catch (TeaException e) {
-            e.printStackTrace();
+            DecryptResponse decryptResponse = client.decryptWithOptions(decryptRequest, runtimeOptions);
+            byte[] plaintext = base64.decode(decryptResponse.getBody().getPlaintext());
+
+            Cipher cipher = Cipher.getInstance("AES/GCM/NoPadding");
+            SecretKeySpec keySpec = new SecretKeySpec(plaintext, "AES");
+            GCMParameterSpec gcmParameterSpec = new GCMParameterSpec(GCM_TAG_LENGTH * 8, iv);
+            cipher.init(Cipher.DECRYPT_MODE, keySpec, gcmParameterSpec);
+            byte[] decryptedData = cipher.doFinal(cipherText);
+            System.out.println(new String(decryptedData));
         } catch (Exception e) {
             e.printStackTrace();
         }
+    }
+
+
+    private static EnvelopeEncryptSample.EnvelopeCipherPersistObject getEnvelopeCipherPersistObject() {
+        // TODO 用户需要在此处代码进行替换，从存储中读取封信加密持久化对象
+        return new EnvelopeEncryptSample.EnvelopeCipherPersistObject();
     }
 
 }
