+++ examples/src/main/java/com/aliyun/kms/kms20160120/samples/GetPublicKeySample.java
@@ -1,31 +1,25 @@
-package com.aliyun.kms.kms20160120;
+package com.aliyun.kms.kms20160120.samples;
 
 import com.aliyun.kms.kms20160120.Client;
 import com.aliyun.kms.kms20160120.model.KmsConfig;
 import com.aliyun.kms.kms20160120.model.KmsRuntimeOptions;
-import com.aliyun.kms20160120.models.DecryptRequest;
-import com.aliyun.kms20160120.models.DecryptResponse;
-import org.apache.commons.codec.binary.Base64;
+import com.aliyun.kms20160120.models.GetPublicKeyRequest;
+import com.aliyun.kms20160120.models.GetPublicKeyResponse;
+import com.aliyun.tea.TeaException;
 
-import javax.crypto.Cipher;
-import javax.crypto.spec.GCMParameterSpec;
-import javax.crypto.spec.SecretKeySpec;
 
-public class EnvelopeDecryptSample {
-    private static Base64 base64 = new Base64();
-
-    private static final int GCM_TAG_LENGTH = 16;
+public class GetPublicKeySample {
 
     public static void main(String[] args) throws Exception {
-        newUserEnvelopeDecrypt();
-        beforeMigrateEnvelopeDecrypt();
-        afterMigrateEnvelopeDecrypt();
+        newUserGetPublicKey();
+        beforeMigrateGetPublicKey();
+        afterMigrateGetPublicKey();
     }
 
     /**
      * 新接入用户可以参考此方法调用KMS实例网关的服务。
      */
-    private static void newUserEnvelopeDecrypt() throws Exception {
+    private static void newUserGetPublicKey() throws Exception {
         //创建kms config并设置相应参数
         com.aliyun.dkms.gcs.openapi.models.Config config
                 = new KmsConfig()
@@ -38,13 +32,13 @@ public class EnvelopeDecryptSample {
                 //设置client-key密码
                 .setPassword("your-password")
                 //设置dkms域名
-                .setEndpoint("your-dkms-endpoint");
-        // 如需验证服务端证书，这里需要设置为您的服务端证书路径
-        //.setCaFilePath("path/to/yourCaCert")
+                .setEndpoint("your-dkms-endpoint")
+                // 如需验证服务端证书，这里需要设置为您的服务端证书路径
+                .setCaFilePath("path/to/yourCaCert");
         // 或者，设置为您的服务端证书内容
         //.setCa("your-ca-certificate-content"));
         Client client = new Client(config);
-        envelopeDecrypt(client);
+        getPublicKey(client);
     }
 
     /**
@@ -52,7 +46,7 @@ public class EnvelopeDecryptSample {
      *
      * @throws Exception
      */
-    private static void beforeMigrateEnvelopeDecrypt() throws Exception {
+    private static void beforeMigrateGetPublicKey() throws Exception {
         //创建kms config并设置相应参数
         com.aliyun.teaopenapi.models.Config config = new com.aliyun.teaopenapi.models.Config()
                 //设置KMS共享网关的域名
@@ -63,7 +57,7 @@ public class EnvelopeDecryptSample {
                 .setAccessKeySecret(System.getenv("your-sk-env-name"));
         //创建kms client
         Client client = new Client(config);
-        envelopeDecrypt(client);
+        getPublicKey(client);
     }
 
     /**
@@ -71,7 +65,7 @@ public class EnvelopeDecryptSample {
      *
      * @throws Exception
      */
-    private static void afterMigrateEnvelopeDecrypt() throws Exception {
+    private static void afterMigrateGetPublicKey() throws Exception {
         //创建kms config并设置相应参数
         com.aliyun.teaopenapi.models.Config config = new com.aliyun.teaopenapi.models.Config()                //设置KMS共享网关的域名
                 .setEndpoint("your-kms-endpoint")
@@ -91,53 +85,37 @@ public class EnvelopeDecryptSample {
                 //设置client-key密码
                 .setPassword("your-password")
                 //设置dkms域名
-                .setEndpoint("your-dkms-endpoint");
-        // 如需验证服务端证书，这里需要设置为您的服务端证书路径
-        //.setCaFilePath("path/to/yourCaCert")
+                .setEndpoint("your-dkms-endpoint")
+                // 如需验证服务端证书，这里需要设置为您的服务端证书路径
+                .setCaFilePath("path/to/yourCaCert");
         // 或者，设置为您的服务端证书内容
         //.setCa("your-ca-certificate-content"));
         //创建kms client
         Client client = new Client(config, kmsConfig);
-        envelopeDecrypt(client);
+        getPublicKey(client);
     }
 
     /**
-     * 信封解密
-     *
-     * @param client
-     * @throws Exception
+     * 获取公钥
      */
-    private static void envelopeDecrypt(Client client) throws Exception {
-
-        EnvelopeEncryptSample.EnvelopeCipherPersistObject envelopeCipherPersistObject = getEnvelopeCipherPersistObject();
-        String encryptedDataKey = envelopeCipherPersistObject.getEncryptedDataKey();
-        byte[] iv = base64.decode(envelopeCipherPersistObject.getIv());
-        byte[] cipherText = base64.decode(envelopeCipherPersistObject.getCipherText());
+    public static void getPublicKey(Client client) {
 
         try {
-            DecryptRequest decryptRequest = new DecryptRequest();
-            decryptRequest.setCiphertextBlob(encryptedDataKey);
+            GetPublicKeyRequest request = new GetPublicKeyRequest();
+            request.setKeyId("your-key-id");
+            request.setKeyVersionId("your-key-version-id");
             KmsRuntimeOptions runtimeOptions = new KmsRuntimeOptions();
             //如需忽略SSL证书认证,可打开如下代码并设置ignoreSSLVerifySwitch为true
             //runtimeOptions.ignoreSSL = true;
-            DecryptResponse decryptResponse = client.decryptWithOptions(decryptRequest, runtimeOptions);
-            byte[] plaintext = base64.decode(decryptResponse.getBody().getPlaintext());
-
-            Cipher cipher = Cipher.getInstance("AES/GCM/NoPadding");
-            SecretKeySpec keySpec = new SecretKeySpec(plaintext, "AES");
-            GCMParameterSpec gcmParameterSpec = new GCMParameterSpec(GCM_TAG_LENGTH * 8, iv);
-            cipher.init(Cipher.DECRYPT_MODE, keySpec, gcmParameterSpec);
-            byte[] decryptedData = cipher.doFinal(cipherText);
-            System.out.println(new String(decryptedData));
+            GetPublicKeyResponse response = client.getPublicKeyWithOptions(request, runtimeOptions);
+            System.out.printf("KeyId: %s%n", response.getBody().getKeyId());
+            System.out.printf("KeyVersionId: %s%n", response.getBody().getKeyVersionId());
+            System.out.printf("PublicKey: %s%n", response.getBody().getPublicKey());
+            System.out.printf("RequestId: %s%n", response.getBody().getRequestId());
+        } catch (TeaException e) {
+            e.printStackTrace();
         } catch (Exception e) {
             e.printStackTrace();
         }
     }
-
-
-    private static EnvelopeEncryptSample.EnvelopeCipherPersistObject getEnvelopeCipherPersistObject() {
-        // TODO 用户需要在此处代码进行替换，从存储中读取封信加密持久化对象
-        return new EnvelopeEncryptSample.EnvelopeCipherPersistObject();
-    }
-
 }
